% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_data.R
\docType{data}
\name{spe_data}
\alias{spe_data}
\title{Phytoplankton species data provided by ocean monitoring in Ecuador}
\format{
\subsection{\code{spe_data}}{

A data frame with 44 rows and 23 columns \if{html}{\out{<species by gender>}}:
\describe{
\item{e1}{affinis \if{html}{\out{<Chaetoceros>}}}
\item{e2}{curvisetus \if{html}{\out{<Chaetoceros>}}}
\item{e3}{didymus \if{html}{\out{<Chaetoceros>}}}
\item{e4}{sinensis \if{html}{\out{<Hemiaulus>}}}
\item{e5}{danicus \if{html}{\out{<Leptocylindrus>}}}
\item{e6}{brightwellii \if{html}{\out{<Ditylum>}}}
\item{e7}{fragilissimus \if{html}{\out{<Dactyliosolen>}}}
\item{e8}{striata \if{html}{\out{<Guinardia>}}}
\item{e9}{imbricata \if{html}{\out{<Rhizosolenia>}}}
\item{e10}{alata \if{html}{\out{<Proboscia>}}}
\item{e11}{antarcticus \if{html}{\out{<Dactyliosolen>}}}
\item{e12}{costatum \if{html}{\out{<Skeletonema>}}}
\item{e13}{borealis \if{html}{\out{<Lauderia>}}}
\item{e14}{sp. \if{html}{\out{<Thalassiosira>}}}
\item{e15}{sp. \if{html}{\out{<Nitzschia>}}}
\item{e16}{longissima \if{html}{\out{<Nitzschia>}}}
\item{e17}{pungens \if{html}{\out{<Pseudo-nitzschia>}}}
\item{e18}{sp. \if{html}{\out{<Navicula>}}}
\item{e19}{membranacea \if{html}{\out{<Stauroneis>}}}
\item{e20}{nitzschioides \if{html}{\out{<Thalassionema>}}}
\item{e21}{sp. \if{html}{\out{<Gymnodinium>}}}
\item{e22}{sp. \if{html}{\out{<Gyrodinium>}}}
\item{e23}{rubrum \if{html}{\out{<Mesodinium>}}}
}
}
}
\source{
Oceanographic Institute of the Navy of Ecuador (INOCAR)
}
\usage{
spe_data
}
\description{
Measurement of the abundance \if{html}{\out{<cells L-1>}} of 23 phytoplankton species
from sampling data at four stations located 10 miles from
the coastal profile of the continental equator.
Data sorted according to \link{env_data} and \link{etiq_data} rows.
}
\keyword{datasets}
