% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumaguayaca.R
\name{sumaguayaca}
\alias{sumaguayaca}
\title{Función suma guayaca}
\usage{
sumaguayaca(a, b, c, d)
}
\arguments{
\item{a}{primer numero}

\item{b}{segundo numero}

\item{c}{tercer numero}

\item{d}{cuarto numero}
}
\value{
el resultado de la suma guayaca
}
\description{
Función suma guayaca
}
\examples{
sumaguayaca(a = 1, b = 2, c = 3, d = 4)
}
